pyproject.toml
[tool.poetry]
name = "chivalry2-unofficial-server-browser-discord-bot"
version = "0.2.0"
description = "A server browser backend for unofficial chivalry2 servers. Keeps track of live servers with a heartbeat, and allows clients to list them. Probably usable for other unofficial gameservers."
authors = ["Jacob Barber <Jacoby6000@gmail.com>"]
license = "GPL3"
readme = "README.md"
packages = [{include = "server_browser_discord_bot", from="src/"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.1"
argparse = "^1.4.0"
requests = "^2.31.0"
tabulate = {extras = ["widechars"], version = "^0.9.0"}
types-requests = "^2.31.0.2"
types-tabulate = "^0.9.0.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
autoflake = "^2.2.0"
mypy = "^1.4.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [ "src" ]
